4.15-4.23 (#5)

+++ part4/controllers/blogs.js
@@ -2,23 +2,43 @@ const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 
 blogsRouter.get("/", async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog
+    .find({}).populate("user", { username: 1, name: 1 })
   res.json(blogs);
 });
 
 blogsRouter.post("/", async (req, res) => {
-  const blog = new Blog(req.body);
-  if (!blog.title || !blog.url) {
+  const { title, author, url, likes } = req.body;
+  const user = req.user;
+
+  if (!title || !url) {
     res.status(400).end();
   } else {
+    const blog = new Blog({
+      title,
+      author,
+      url,
+      likes,
+      user: user.id,
+    });
+
     const savedBlog = await blog.save();
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
     res.status(201).json(savedBlog);
   }
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
-  await Blog.findByIdAndDelete(req.params.id);
-  res.status(204).end();
+  const user = req.user;
+  const blog = await Blog.findById(req.params.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(req.params.id);
+    res.status(204).end();
+  } else {
+    return res.status(401).json({ error: "unauthorized" });
+  }
 });
 
 blogsRouter.put("/:id", async(req, res) => {
@@ -29,7 +49,7 @@ blogsRouter.put("/:id", async(req, res) => {
     author,
     url,
     likes,
-  };
+  }
 
   await Blog.findByIdAndUpdate(req.params.id, updatedBlog);
   res.status(200).end();
